Refactor `parseRemoteFileAddress` out of `getRemoteConfig`
